Refactor unit tests and add test for same successive intents
