+++ test/SingleSessionChatbotTest.cpp
@@ -44,7 +44,7 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
 
-#include "../launcher/TestContext.hpp"
+#include "launcher/TestContext.hpp"
 #include "json.hpp"
 
 #include "intent/utils/Deserializer.hpp"
@@ -65,7 +65,7 @@ namespace intent
             {
                 const intent::test::ResourceManager &resourceManager = intent::test::gTestContext->getResourceManager();
 
-                std::string jsonContent = resourceManager.getResource(test::ResourceManager::ResourceId::CHATBOT_MODE_JSON);
+                std::string jsonContent = resourceManager.getResource(test::ResourceManager::ResourceId::CHATBOT_MODEL_JSON);
 
                 nlohmann::json json = nlohmann::json::parse(jsonContent);
