+++ test/same_successive_intents/res/user_commands.js
@@ -1,0 +1,3 @@
+
+module.exports = {}
+
